feat: Add ability to configure custom comparsions for specific members.
This feature adds the ability to use the configuration context to
configure any member to be compared with a custom IComparison
implementation.  The use of the IComparison interface was originally
introduced to support property-level comparison strategies with
anonymous objects.  This feature extends this capability to comparisons
with any object.  In conjunction with this feature, a new
IgnoredComparision implementation has been added to support the ability
to ignore comparisons for specific members.